+++ FrontEnd/src/components/NavBar.jsx
@@ -15,9 +15,9 @@ function NavBar() {
                 <Navbar.Toggle aria-controls="basic-navbar-nav" />
                 <Navbar.Collapse id="basic-navbar-nav">
                     <Nav className="me-auto">
-                        <NavLink to="/ProfilePage" className="navbaritem body px-3">Profile</NavLink>
+                        <NavLink to="/ProfilePage" className="navbaritem body px-3">Home</NavLink>
                         {/* using NavLinks here, when using bootstrap Nav.Link the context was lost when navigation occurred */}
-                        <NavLink to="/TimelinePage" className="navbaritem body px-3">Timeline</NavLink>
+                        <NavLink to="/TimelinePage" className="navbaritem body px-3">Browse</NavLink>
                         <NavLink to="/MessagesPage" className="navbaritem body px-3">Messages</NavLink>
                         <NavLink to="/SavedItemsPage" className="navbaritem body px-3">Saved</NavLink>
                         <NavLink to="/NotificationsPage" className="navbaritem body px-3">Notifications</NavLink>
