+++ FrontEnd/src/pages/ProfilePage.jsx
@@ -6,6 +6,8 @@ import NavBar from '../components/NavBar';
 import LogOutButton from '../components/LogOutButton';
 import RentedItemCard from '../components/RentedItemCard';
 import SavedItemsCard from '../components/SavedItemsCard';
+import useFetch from '../hooks/UseFetch';
+
 // import the relevant bootstrap:
 import Col from 'react-bootstrap/esm/Col';
 import Row from 'react-bootstrap/Row';
@@ -14,40 +16,40 @@ import Container from 'react-bootstrap/esm/Container';
 import { useUserContext } from '../context/userContext'
 import { useEffect, useState } from 'react';
 
-const useFetch = (url) => {
-    const [data, setData] = useState([]);
-    const [error, setError] = useState(null);
+// const useFetch = (url) => {
+//     const [data, setData] = useState([]);
+//     const [error, setError] = useState(null);
 
-    useEffect(() => {
-        const fetchData = async () => {
-            try {
-                const response = await fetch(url);
-                const result = await response.json();
+//     useEffect(() => {
+//         const fetchData = async () => {
+//             try {
+//                 const response = await fetch(url);
+//                 const result = await response.json();
 
-                if (response.ok) {
-                    setData(result.data);
-                } else {
-                    setError(result.error);
-                }
-            } catch (error) {
-                setError(error.message);
-            }
-        };
+//                 if (response.ok) {
+//                     setData(result.data);
+//                 } else {
+//                     setError(result.error);
+//                 }
+//             } catch (error) {
+//                 setError(error.message);
+//             }
+//         };
 
-        fetchData();
-    }, [url]);
+//         fetchData();
+//     }, [url]);
 
-    return { data, error };
-};
+//     return { data, error };
+// };
 
 
 function ProfilePage() {
     const currentUser = useUserContext();
     const userItemsUrl = `http://localhost:3307/rentshare/items/getrented/${parseInt(currentUser.currentUser.UserID, 10)}`;
-    const userSavedItemsUrl = `http://localhost:3307/rentshare/saveditems/getsaveddetails/${parseInt(currentUser.currentUser.UserID, 10)}`;
+    // const userSavedItemsUrl = `http://localhost:3307/rentshare/saveditems/getsaveddetails/${parseInt(currentUser.currentUser.UserID, 10)}`;
 
     const { data: userItems, error: errorItems } = useFetch(userItemsUrl);
-    const { data: userSavedItems, error: errorSavedItems } = useFetch(userSavedItemsUrl);
+    // const { data: userSavedItems, error: errorSavedItems } = useFetch(userSavedItemsUrl);
 
     if (!currentUser.currentUser.UserID) {
         return <div>Loading...</div>; // Or any loading indicator
@@ -77,14 +79,14 @@ function ProfilePage() {
                                     <div>No current items for rent</div>
                                 )}
                             </Col>
-                            <Col md={6}>
+                            {/* <Col md={6}>
                                 <h3 className='headings' style={{marginBottom: '55px'}}>Items you have saved:</h3>
                                 {userSavedItems.length > 0 ? (
                                     <SavedItemsCard savedItems={userSavedItems} />
                                 ) : (
                                     <div>No saved items</div>
                                 )}
-                            </Col>
+                            </Col> */}
                         </Row>
                     </Col>
                 </Row>
